+++ src/genome/tf/motif/__init__.py
@@ -12,6 +12,8 @@ from genome.tf.motif.jaspar import (
 )
 from genome.tf.motif.motif import (
     BASES,
+    DEFAULT_SEQUENCE_NAME,
+    DEFAULT_THRESHOLD,
     MIN_MOTIF_LENGTH,
     AmbiguousBaseIdError,
     AmbiguousMotifNameError,
@@ -19,16 +21,28 @@ from genome.tf.motif.motif import (
     MotifNotFoundError,
     MotifSet,
 )
+from genome.tf.motif.scan import (
+    HIT_COLUMNS,
+    HIT_DTYPES,
+    HIT_PROVENANCE,
+    FastaFormatError,
+)
 
 __all__ = [
     "BASES",
     "DEFAULT_RELEASE",
+    "DEFAULT_SEQUENCE_NAME",
     "DEFAULT_TAX_GROUP",
+    "DEFAULT_THRESHOLD",
+    "HIT_COLUMNS",
+    "HIT_DTYPES",
+    "HIT_PROVENANCE",
     "JASPAR_RELEASES",
     "JASPAR_TAX_GROUPS",
     "MIN_MOTIF_LENGTH",
     "AmbiguousBaseIdError",
     "AmbiguousMotifNameError",
+    "FastaFormatError",
     "JasparDatabase",
     "JasparReleaseError",
     "Motif",